pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "watchflow"
version = "0.1.0"
description = "Production-grade reactive automation CLI platform with semantic intent engine"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "ScorpioCodeX", email = "scorpiocodex0@gmail.com" }]
keywords = ["automation", "file-watcher", "cli", "reactive", "dag", "sci-fi", "next-gen"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Build Tools",
]
dependencies = [
    "typer>=0.12",
    "rich>=13.7",
    "watchdog>=4.0",
    "pydantic>=2.0",
    "structlog>=24.0",
    "anyio>=4.0",
    "uvloop>=0.19; sys_platform != 'win32'",
    "networkx>=3.3",
    "psutil>=5.9",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "hypothesis>=6.100",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-PyYAML",
    "types-psutil",
]

[project.scripts]
watchflow = "watchflow.cli.main:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
src = ["src"]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true
packages = ["watchflow"]

[tool.coverage.run]
source = ["src/watchflow"]
